+++ app/Http/Controllers/EventController.php
@@ -2339,34 +2339,37 @@ class EventController extends Controller
 
     }
     public function OptimizeImages ($url,$filename){
-        $realpath = public_path($url);
-        if(file_exists($realpath.'/'.$filename)){
-            if(!file_exists($realpath.'/origin'))
-        mkdir($realpath.'/origin', 0777, true);
-    
-        if(!file_exists($realpath.'/medium'))
-        mkdir($realpath.'/medium', 0777, true);
-         
+        if($filename!=""){
+            $realpath = public_path($url);
+            if(file_exists($realpath.'/'.$filename)){
+                if(!file_exists($realpath.'/origin'))
+            mkdir($realpath.'/origin', 0777, true);
         
-    
-        copy($realpath.'/'.$filename,$realpath.'/origin/'.$filename);
-        $imagesize = round(filesize($realpath.'/origin/'.$filename)/1024/1024); 
-    
-        if($imagesize<1)
-           copy($realpath.'/'.$filename,$realpath.'/medium/'.$filename);
+            if(!file_exists($realpath.'/medium'))
+            mkdir($realpath.'/medium', 0777, true);
+             
+            
         
-        else{
-            File::delete($realpath.'/'.$filename);
-            list($width, $height, $type, $attr) = getimagesize($realpath.'/origin/'.$filename);
+            copy($realpath.'/'.$filename,$realpath.'/origin/'.$filename);
+            $imagesize = round(filesize($realpath.'/origin/'.$filename)/1024/1024); 
+        
+            if($imagesize<1)
+               copy($realpath.'/'.$filename,$realpath.'/medium/'.$filename);
+            
+            else{
+                File::delete($realpath.'/'.$filename);
+                list($width, $height, $type, $attr) = getimagesize($realpath.'/origin/'.$filename);
+        
+                $image_medium = new \Intervention\Image\ImageManager();
+                $image_medium->make($realpath.'/origin/'.$filename)->resize($width/2,$height/2)->save($realpath.'/medium/'.$filename);
+              
+                $image_small = new \Intervention\Image\ImageManager();
+                $image_small->make($realpath.'/origin/'.$filename)->resize($width/3,$height/3)->save($realpath.'/'.$filename);
+            }  
     
-            $image_medium = new \Intervention\Image\ImageManager();
-            $image_medium->make($realpath.'/origin/'.$filename)->resize($width/2,$height/2)->save($realpath.'/medium/'.$filename);
-          
-            $image_small = new \Intervention\Image\ImageManager();
-            $image_small->make($realpath.'/origin/'.$filename)->resize($width/3,$height/3)->save($realpath.'/'.$filename);
-        }  
-
+            }
         }
+
